% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootEta.R
\name{bootEta}
\alias{bootEta}
\title{Bootstrap samples of profiled eta}
\usage{
bootEta(y, B = 1000, method = "P")
}
\arguments{
\item{y}{Observation matrix. n by p.}

\item{B}{Number of samples. Default B=1000.}

\item{method}{String. Parametric (P) or no parametric (NP)}
}
\description{
Bootstrap samples of profiled eta
}
