% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{delta}
\alias{delta}
\alias{pdelta.slash}
\alias{qdelta.slash}
\alias{ddelta.slash}
\title{Distribution, quantile and density of delta}
\usage{
pdelta.slash(x, eta, p)

qdelta.slash(qtl, eta, p)

ddelta.slash(x, eta, p)
}
\arguments{
\item{x}{Value of \emph{delta}.}

\item{eta}{Parameter between 0 and 1.}

\item{p}{Dimension of \code{Y}.}

\item{qtl}{Quantile value. Between 0 and 1.}
}
\value{
A distribution, density or quantile value.
}
\description{
Obtain distribution, quantile or density.
}
\details{
Evaluate analitic expected values.
}
\examples{
n <- 1000
eta <- 1/3
p <- 4
Z <- sim.slash( n, rep(0,p), diag(p), eta )
delta <- sapply(1:n, function(i) Z[i,,drop=FALSE]\%*\%t(Z[i,,drop=FALSE]) )

hist(delta, prob=TRUE)
curve(ddelta.slash(x, eta, p), add=TRUE, col=2)

}
