% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{dgfg}
\alias{dgfg}
\alias{dg.slash}
\alias{fg.slash}
\title{Evaluate integrals \emph{dg} and \emph{fg}.}
\usage{
dg.slash(eta, p)

fg.slash(eta, p)
}
\arguments{
\item{eta}{Parameter between 0 and 1.}

\item{p}{Dimension of \code{Y}.}

\item{x}{Value of \emph{delta}.}

\item{qtl}{Quantile value. Between 0 and 1.}
}
\value{
A \code{numeric}.
}
\description{
Obtain \emph{dg} and \emph{fg} values from numerical integrals.
}
\details{
Evaluate integrals \emph{dg} and \emph{fg}.
}
\examples{
p <- 4
etas <- seq(1/50, 1-1/50, length.out=200)
dgs <- sapply( etas, function(eta) dg.slash( eta, p ) )
fgs <- sapply( etas, function(eta) fg.slash( eta, p ) )

plot( etas, dgs, xlim=c(0,1), type="l")
abline( h=p/4, col=2 )

plot( etas, fgs, xlim=c(0,1), type="l")
abline( h=p*(p+2)/4, col=2 )

}
