% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{em.slash}
\alias{em.slash}
\title{EM algotithm}
\usage{
em.slash(Y, mu0, Sigma0, eta0, relError = 1e-05, verbose = FALSE)
}
\arguments{
\item{Y}{Sample slash.}

\item{mu0}{Mean vector with length \code{p}.}

\item{Sigma0}{Covariance matrix with size \code{p}x\code{p}.}

\item{eta0}{Parameter between 0 and 1.}

\item{relError}{Relative error.}

\item{verbose}{Logical. \code{TRUE} indicate trace the steps.}
}
\value{
A \code{list} with \code{mu0}, \code{Sigma0} and \code{eta0} components.
}
\description{
Obtain MLE values from EM algoritm.
}
\details{
Conditioning \emph{v} in step E. Use \code{startValues.slash} function to obtain start values
}
\examples{
n <- 1000
p <- 4
Y <- sim.slash( n, rep(0, p), diag(p), 1/2 )
sv <- startValues.slash(Y)
em.slash( Y, sv$mu0 , sv$Sigma0, sv$eta0, verbose=TRUE )

}
