% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopeForCalpha.R
\name{envelopeForCalpha}
\alias{envelopeForCalpha}
\title{Envelopes for C(alpha)}
\usage{
envelopeForCalpha(y, distr = "mvnorm", method = "moment", reps = 20,
  conf = 0.95, ...)
}
\description{
QQplot with envelopes
}
