% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoFisher.R
\name{infoFisher.slash}
\alias{infoFisher.slash}
\title{Fisher Information Matrix Slash}
\usage{
infoFisher.slash(Y, mu, Sigma, eta, block = TRUE)
}
\arguments{
\item{Y}{Sample slash.}

\item{mu}{Mean vector with length \code{p}.}

\item{Sigma}{Covariance matrix with size \code{p}x\code{p}.}

\item{eta}{Parameter between 0 and 1.}

\item{block}{Returns a blocked matrix.}
}
\value{
A \code{list} with \code{I.mu.mu}, \code{I.phi.phi} and \code{U.eta} components.
}
\description{
Obtain a Information matrix values for parameters.
}
\details{
Evaluate analitic expected values.
}
\examples{
n <- 1000
p <- 4
Y <- sim.slash( n, rep(0, p), diag(p), 1/2 )
sv <- startValues.slash(Y)
fit <- em.slash( Y, sv$mu0 , sv$Sigma0, sv$eta0, verbose=FALSE )
infoFisher.slash(Y, fit$mu, fit$Sigma, fit$eta)

}
