% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappaEst.R
\name{kappaEst}
\alias{kappaEst}
\title{Multivariate excess of kurtosis}
\usage{
kappaEst(y, eta = NULL)
}
\arguments{
\item{y}{Matrix n rows by p columns.}

\item{eta}{Numeric. Default NULL}
}
\value{
A numeric value
}
\description{
Evaluate multivariate excess of kurtosis according emphirical kappa
or expected kappa.
}
