% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testLMforGARCH.R
\name{LMgarch.test}
\alias{LMgarch.test}
\title{Test Lagrange Multipliers}
\usage{
LMgarch.test(y, p, q)
}
\arguments{
\item{y}{Numeric vector}

\item{p}{ARCH order}

\item{q}{GARCH order}
}
\value{
A list
\describe{
  \item{LM}{Statistic}
  \item{p.value}{p value}
  \item{df}{Degree of freedom}
}
}
\description{
Test Lagrange Multipliers
}
\examples{
set.seed(1984)
y <- rnorm(500)
LMgarch.test(y, 1, 1)
# H0: alpha_1=beta_1=0 
# is accepted
}
\references{
Lee, J. 1991. A Lagrange multiplier test for GARCH models. Economics Letters 37: 265-271.
}
