% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{loglik.slash}
\alias{loglik.slash}
\title{Log-likelihood of multivariate slash reparemetrized}
\usage{
loglik.slash(Y, mu, Sigma, eta)
}
\arguments{
\item{Y}{Sample slash.}

\item{mu}{Mean vector with length \code{p}.}

\item{Sigma}{Covarianze matrix with size \code{p}x\code{p}.}

\item{eta}{Parameter between 0 and 1.}
}
\value{
A \code{numeric}.
}
\description{
Obtain a log-likelihood of a sample Slash.
}
\details{
Compute the log-likelihood without conditioning.
}
\examples{
n <- 1000
p <- 4
Y <- sim.slash( n, rep(0, p), diag(p), 1/2 )
loglik.slash(Y, rep(0, p), diag(p), 1/2 )

}
