% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmegaMat.R
\name{OmegaMat}
\alias{OmegaMat}
\title{Asymptotic covariance Omega}
\usage{
OmegaMat(Sigma, eta, kappa = NULL)
}
\arguments{
\item{Sigma}{Covariance matrix estimator.}

\item{eta}{Shape parameter estimator.}

\item{kappa}{Multivariate excess of kurtosis estimator.}
}
\value{
A square matrix of p + p*(p+1)/2
}
\description{
Evaluate asymptotic covariance of
}
