% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{qqplot.slash}
\alias{qqplot.slash}
\title{qq-plot Slash}
\usage{
qqplot.slash(Y, mu, Sigma, eta)
}
\arguments{
\item{Y}{Sample slash.}

\item{mu}{Mean vector with length \code{p}.}

\item{Sigma}{Covariance matrix with size \code{p}x\code{p}.}

\item{eta}{Parameter between 0 and 1.}
}
\value{
A qq-plot.
}
\description{
Obtain a qq-plot for \emph{delta}.
}
\details{
Use distribution of \emph{delta} values
}
\examples{
n <- 1000
p <- 4
Y <- sim.slash( n, rep(0, p), diag(p), 1/2 )
sv <- startValues.slash(Y)
fit <- em.slash( Y, sv$mu0 , sv$Sigma0, sv$eta0, verbose=FALSE )
qqplot.slash(Y, fit$mu, fit$Sigma, fit$eta)

}
