% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hessian.R
\name{score.slash}
\alias{score.slash}
\title{Score Slash}
\usage{
score.slash(Y, mu, Sigma, eta, block = TRUE, hessian = FALSE)
}
\arguments{
\item{Y}{Sample slash.}

\item{mu}{Mean vector with length \code{p}.}

\item{Sigma}{Covariance matrix with size \code{p}x\code{p}.}

\item{eta}{Parameter between 0 and 1.}

\item{block}{Returns a blocked score.}

\item{hessian}{Logical. Is the hessian matrix evaluated?}
}
\value{
A \code{list} with \code{U.mu}, \code{U.Sigma} and \code{U.eta} components if \code{block=FALSE} and \code{hessian=FALSE}
}
\description{
Obtain a score values for parameters.
}
\details{
Evaluate analitic loglik derivatives.
}
\examples{
n <- 1000
p <- 4
Y <- sim.slash( n, rep(0, p), diag(p), 1/2 )
sv <- startValues.slash(Y)
fit <- em.slash( Y, sv$mu0 , sv$Sigma0, sv$eta0, verbose=FALSE )
qqplot.slash(Y, fit$mu, fit$Sigma, fit$eta)
loglik.slash(Y, fit$mu , fit$Sigma, fit$eta )
score.slash(Y, fit$mu , fit$Sigma, fit$eta )

}
