% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_slash.R
\name{sim.slash}
\alias{sim.slash}
\title{Simulate multivariate slash reparemetrized}
\usage{
sim.slash(n, mu, Sigma, eta)
}
\arguments{
\item{n}{Number of simulates}

\item{mu}{Mean vector with length \code{p}.}

\item{Sigma}{Covarianze matrix with size \code{p}x\code{p}.}

\item{eta}{Parameter between 0 and 1.}
}
\value{
A matrix with \code{n} rows and \code{p} .
}
\description{
Return a multivariate slash simulation.
}
\details{
Let v distruted beta with parameter 1/eta and 1. And Z distributed multivariate normal
with mean zero and  covariance matrix \emph{Sigma}. We obtain a sample slash \emph{Y} with
Y = mu + sqrt(1-eta)/sqrt(v) Z
}
\examples{
n <- 1000
p <- 4
Y <- sim.slash( n, rep(0, p), diag(p), 1/2 )
plot(Y, asp=1)

}
