% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\docType{package}
\name{slash}
\alias{slash}
\alias{slash-package}
\title{Slash reparametrized}
\description{
The package \pkg{slash} implements routines to EM estimate.
}
\details{
Several utilities for ML estimators.
}
