% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{startValues.slash}
\alias{startValues.slash}
\title{Start values for EM algotithm}
\usage{
startValues.slash(Y, mu0 = NULL, Sigma0 = NULL, eta0 = NULL)
}
\arguments{
\item{Y}{Sample slash.}

\item{mu0}{Specify mu0. Default is \code{NULL}.}
}
\value{
A \code{list} with \code{mu0}, \code{Sigma0} and \code{eta0} components.
}
\description{
Obtain start values for EM algoritm plug-in estimators of mean and covariance matrix. \emph{eta} is 
calculate from coeficient of multivariate kurtosis
}
\details{
Compute Start values for EM algotithm.
}
\examples{
n <- 1000
p <- 4
Y <- sim.slash( n, rep(0, p), diag(p), 1/2 )
startValues.slash(Y)

}
