% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_alpha_twoGroup.R
\name{twoGroups.slash}
\alias{twoGroups.slash}
\title{Evaluate C(alpha) for group comparation.}
\usage{
twoGroups.slash(g1, g2)
}
\arguments{
\item{g1}{matrix (n1 rows and p columns).}

\item{g2}{matrix (n2 rows and p columns).}
}
\value{
A \code{list} with \code{C.alpha}, \code{p.value}, \code{k1} and estimated values \code{sv}.
}
\description{
Evaluate C(alpha) for two group comparation.
}
\details{
C(alpha) is based in sqrt(n)-consistent estimators.
}
\examples{
p <- 5
g1 <- sim.slash(100, rep(0,p), diag(p), 0.45)
g2 <- sim.slash(100, rep(1,p), diag(p), 0.45)
twoGroups.slash(g1, g2)

g1 <- sim.slash(100, rep(0,p), diag(p), 0.45)
g2 <- sim.slash(100, rep(0,p), diag(p), 0.45)
twoGroups.slash(g1, g2)

}
